fix: CFD-306 Fix issue with shared types (#56)
