+++ repos/fdbt-netex-output/src/netex-convertor/point-to-point-tickets/pointToPointTicketNetexHelpers.ts
@@ -1,5 +1,4 @@
 import startCase from 'lodash/startCase';
-import { SalesOfferPackage } from '../../../../../shared/matchingJsonTypes';
 import {
     FareZone,
     FareZoneList,
@@ -8,6 +7,7 @@ import {
     PointToPointTicket,
     DistributionAssignment,
     SalesOfferPackageElement,
+    SalesOfferPackage,
     BaseProduct,
     NetexSalesOfferPackage,
     isReturnTicket,
