+++ repos/fdbt-site/package.json
@@ -141,7 +141,6 @@
         "mock-req-res": "^1.1.4",
         "mock-res": "^0.5.0",
         "prettier": "2.3.2",
-        "pretty-quick": "^3.1.1",
         "sass": "^1.26.5",
         "sinon": "^9.0.1",
         "ts-node": "^8.10.1",
