+++ repos/fdbt-site/tests/pages/api/describeSalesOfferPackage.test.ts
@@ -1,4 +1,3 @@
-import { SalesOfferPackage } from '../../../../../shared/matchingJsonTypes';
 import describeSalesOfferPackage, {
     sopInfoSchema,
     checkUserInput,
@@ -11,6 +10,7 @@ import * as sessionUtils from '../../../src/utils/sessions';
 import * as aurora from '../../../src/data/auroradb';
 import {
     ErrorInfo,
+    SalesOfferPackage,
     SalesOfferPackageInfo,
     SalesOfferPackageInfoWithErrors,
     SalesOfferPackageWithErrors,
