--- shared/matchingJsonTypes.d.ts
@@ -1,9 +1,0 @@
-export interface SalesOfferPackage {
-  id?: string;
-  name: string;
-  description: string;
-  purchaseLocations: string[];
-  paymentMethods: string[];
-  ticketFormats: string[];
-  price?: string;
-}
