--- shared/tsconfig.json
@@ -1,4 +1,0 @@
-{
-  "extends": "../repos/fdbt-site",
-  "include": ["./**"]
-}
